apis/pyproject.toml
[tool.poetry]
name = "robin-api"
version = "0.1.0"
description = "Multilingual product API with FastAPI + Strawberry"
authors = ["Toi <toi@exemple.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
strawberry-graphql = "^0.222.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0"
pydantic = "^2.6"
mysqlclient = "^2.2"
python-dotenv = "^1.0.1"
transformers = "^4.39.3"
torch = "^2.2.2"
huggingface_hub = {extras = ["hf_xet"], version = "^0.23.1"}

[tool.poetry.group.dev.dependencies]
rich = "^14.0.0"
typer = "^0.15.2"
libcst = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
